WIP: Applying feedback
